backend/requirements.txt
fastapi
uvicorn
pandas
db-dtypes
google-cloud-storage
google-cloud-bigquery